requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
oauth2client
httplib2
tqdm